requirements.txt
# LangGraph / LangChain core
langgraph>=0.2.39
langgraph-cli[inmem]>=0.1.46
langchain-community>=0.3.0
langchain-openai>=0.2.3
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0

# Tavily (web search)
tavily-python>=0.3.0

# RAG/DB/Utils
faiss-cpu>=1.8.0
pymysql>=1.1.1
python-dotenv>=1.0.1
pytz>=2024.1
psycopg2-binary>=2.9.9  # PostgreSQL driver for pgvector
langchain-postgres>=0.0.6  # LangChain PGVector integration

# ML Models
transformers>=4.30.0
torch>=2.0.0

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Storage (optional - Redis 사용 시)
# redis>=5.0.0